+++ src/view/sort-button-view.js
@@ -1,5 +1,5 @@
 import AbstractView from './abstract-view.js';
-import {SortType} from '../utils/helpers.js';
+import {SortType} from '../utils/const.js';
 
 const createSortButtonsTemplate = (currentSortType) => (
   `<ul class="sort">
